refactor: User 필드에 Embedded 사용
1. Name, Password, Email not null
2. Name, Password, Email @Embedded 애너테이션 추가
3. Converter에 @Deprecated 추가
