+++ src/main/java/techcourse/myblog/domain/user/UserPassword.java
@@ -1,13 +1,18 @@
 package techcourse.myblog.domain.user;
 
+import javax.persistence.Embeddable;
 import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+@Embeddable
 public class UserPassword {
     private static final Pattern PASSWORD_PATTERN = Pattern.compile("^(?=.*[A-Z])(?=.*[a-z])(?=.*[0-9])(?=.*[$@$!%*#?&])[A-Za-z[0-9]$@$!%*#?&]{8,}$");
     private static final String PASSWORD_ERROR = "대소문자, 특수문자, 숫자를 포함한 8자리 이상 문자를 입력하세요. 싫으면 시집가시고";
     private String password;
+
+    private UserPassword() {
+    }
 
     public UserPassword(String password) {
         this.password = validate(password);
