+++ src/main/java/techcourse/myblog/domain/user/converter/UserEmailConverter.java
@@ -1,8 +1,11 @@
-package techcourse.myblog.domain.user;
+package techcourse.myblog.domain.user.converter;
+
+import techcourse.myblog.domain.user.UserEmail;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
+@Deprecated
 @Converter
 public class UserEmailConverter implements AttributeConverter<UserEmail, String> {
     @Override
