+++ src/main/java/techcourse/myblog/domain/user/converter/UserNameConverter.java
@@ -1,8 +1,11 @@
-package techcourse.myblog.domain.user;
+package techcourse.myblog.domain.user.converter;
+
+import techcourse.myblog.domain.user.UserName;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
+@Deprecated
 @Converter
 public class UserNameConverter implements AttributeConverter<UserName, String> {
     @Override
