+++ src/main/java/techcourse/myblog/domain/user/converter/UserPasswordConverter.java
@@ -1,8 +1,11 @@
-package techcourse.myblog.domain.user;
+package techcourse.myblog.domain.user.converter;
+
+import techcourse.myblog.domain.user.UserPassword;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
+@Deprecated
 @Converter
 public class UserPasswordConverter implements AttributeConverter<UserPassword, String> {
     @Override
